added styling and changed the homepage and the homepage for a user
