+++ react-app/public/images/quickpic-logo.png
