+++ react-app/src/components/LoginForm/LoginForm.css
@@ -1,0 +1,107 @@
+/* .Login-Form{
+
+
+
+
+} */
+
+.login-ENTIRE-container-div{
+    display:flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+}
+.login-container-div{
+    padding: 2px 10px;
+    display: flex;
+    flex-direction: column;
+    margin: 100px;
+    margin-bottom: 15px;
+    border: 1px solid rgba(145, 142, 142, 0.74);
+    border-radius: 5px 5px;
+    justify-content: center;
+    align-items: center;
+
+}
+
+.inputvalues{
+border: 1px solid rgba(95, 95, 95, 0.459);
+margin: 5px;
+border-radius: 5px 5px;
+padding: 10px ;
+padding-right: 200px;
+background-color: rgba(224, 217, 217, 0.582);
+font-family: sans-serif;
+}
+
+.login-button{
+color: white;
+background-color: rgba(83, 206, 223, 0.993);
+border-radius: 5px 5px ;
+padding: 8px 175px;
+border: 1px solid lightblue;
+transition-duration: 1.4s;
+}
+.login-button:hover{
+color: white;
+background-color: rgba(61, 118, 172, 0.993);
+cursor: pointer;
+
+}
+
+.logo-quicksPic{
+    max-width: 400px;
+    margin: 40px;
+}
+
+.Input-values-and-login{
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+    align-items: center;
+}
+
+
+.demoUserButton{
+
+    border: solid 1px #dd81f5;
+    background-color: #a048e4;
+
+    color:white;
+    font-style: bold;
+    cursor: pointer;
+    border-radius: 5px 5px;
+    padding: 8px 20px;
+    margin: 15px;
+    transition-duration: 1.4s;
+    font-family: sans-serif;
+
+}
+.demoUserButton:hover{
+    background-color:  #dd81f5;
+
+}
+.OR-statement{
+    margin-top: 20px;
+    color:grey;
+}
+
+.or-div-container{
+    margin: 20px;
+}
+
+
+.login-demo-user{
+    color:#4daff4;
+    margin-top: 10px;
+    font-family: sans-serif;
+}
+
+.lines-around-or{
+
+/* background-color: rgba(var(--b38,219,219,219),1); */
+    /* background-color: #dbdbdb; */
+    padding-top: 20px;
+    margin-bottom: 200px;
+    color:rgba(59, 57, 57, 0.658);
+}
