--- react-app/src/components/NavBar.js
@@ -1,41 +1,0 @@
-
-import React from 'react';
-import { NavLink } from 'react-router-dom';
-import LogoutButton from './auth/LogoutButton';
-
-const NavBar = () => {
-  return (
-    <div className='navigationbar-Div'>
-
-      <nav>
-        <ul>
-          <li>
-            <NavLink to='/' exact={true} activeClassName='active'>
-              Home
-            </NavLink>
-          </li>
-          <li>
-            <NavLink to='/login' exact={true} activeClassName='active'>
-              Login
-            </NavLink>
-          </li>
-          <li>
-            <NavLink to='/sign-up' exact={true} activeClassName='active'>
-              Sign Up
-            </NavLink>
-          </li>
-          <li>
-            <NavLink to='/users' exact={true} activeClassName='active'>
-              Users
-            </NavLink>
-          </li>
-          <li>
-            <LogoutButton />
-          </li>
-        </ul>
-      </nav>
-    </div>
-  );
-}
-
-export default NavBar;
