+++ react-app/src/components/SignUpForm/SignUpForm.css
@@ -1,0 +1,24 @@
+.sign-up-button {
+    color: white;
+    background-color: rgba(83, 206, 223, 0.993);
+    border-radius: 5px 5px;
+    padding: 8px 167px;
+    border: 1px solid lightblue;
+    transition-duration: 1.4s;
+    margin-left: 5px;
+}
+
+.sign-up-button:hover {
+    color: white;
+    background-color: rgba(61, 118, 172, 0.993);
+    cursor: pointer;
+
+}
+
+.signUP-div-container {
+display: flex;
+flex-direction: column;
+justify-content: center;
+align-items: center;
+
+}
